mannsaathi-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.0
numpy>=1.26.2
pandas>=2.1.0
pillow>=10.1.0
opencv-python>=4.8.1
librosa>=0.10.1
soundfile>=0.12.1
scikit-learn>=1.3.0
nltk>=3.8.1
textblob>=0.17.1
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langdetect>=1.0.9
transformers>=4.36.0
torch>=2.1.0
fer>=22.5.0
structlog>=23.2.0
httpx>=0.25.0
aiofiles>=23.2.0
websockets>=12.0